feat: implement API key middleware for chat routes and update Chat model types

+++ models/Chat.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Chat struct {
-	ID             int64     `json:"id"`
-	ReminderUserID int64     `json:"reminder_user_id"`
+	ID             int       `json:"id"`
+	ReminderUserID int       `json:"reminder_user_id"`
 	Platform       string    `json:"platform"`
 	ChatID         int64     `json:"chat_id"`
 	CreatedAt      time.Time `json:"created_at"`
